requirements.txt
fastapi
uvicorn
pymongo
jwt
pydantic
passlib
python-multipart
one
bcrypt
cryptography
python-jose